services - mvc arch

+++ src/main/java/com/codingshuttle/restspringweb/services/EmployeeService.java
@@ -4,10 +4,15 @@ package com.codingshuttle.restspringweb.services;
 import com.codingshuttle.restspringweb.dto.EmployeeDTO;
 import com.codingshuttle.restspringweb.entities.EmployeeEntity;
 import com.codingshuttle.restspringweb.repositories.EmployeeRepository;
+import org.apache.el.util.ReflectionUtil;
 import org.modelmapper.ModelMapper;
+import org.springframework.util.ReflectionUtils;
 import org.springframework.stereotype.Service;
 
+import java.lang.reflect.Field;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 
@@ -24,11 +29,13 @@ public class EmployeeService {
     private final ModelMapper modelMapper;
 
 
-    public EmployeeDTO getEmployeeById(Long id) {
+    public Optional<EmployeeDTO> getEmployeeById(Long id) {
 
-        EmployeeEntity employeeEntity = employeeRepository.findById(id).orElse(null);
-        ModelMapper mapper = new ModelMapper();
-        return modelMapper.map(employeeEntity, EmployeeDTO.class);
+//        EmployeeEntity employeeEntity = employeeRepository.findById(id).orElse(null);
+//        ModelMapper mapper = new ModelMapper();
+//        return modelMapper.map(employeeEntity, EmployeeDTO.class);
+
+        return employeeRepository.findById(id).map(employeeEntity -> modelMapper.map(employeeEntity, EmployeeDTO.class));
     }
 
     public List<EmployeeDTO> getAllEmployees() {
@@ -57,9 +64,28 @@ public class EmployeeService {
         return modelMapper.map(savedEmployeeEntity, EmployeeDTO.class);
     }
 
-    public void deleteEmployeeById(Long employeeId) {
+    public boolean isExistsByEmployeeId(Long employeeId){
+        return employeeRepository.existsById(employeeId);
+    }
 
+    public boolean deleteEmployeeById(Long employeeId) {
+        boolean exists = isExistsByEmployeeId(employeeId);
+        if(!exists) return false;
         employeeRepository.deleteById(employeeId);
+        return true;
+    }
+
+    public EmployeeDTO updatePartialEmployeeById(Long employeeId, Map<String, Object> updates) {
+        boolean exists = isExistsByEmployeeId(employeeId);
+        if(!exists) return null;
+        EmployeeEntity employeeEntity = employeeRepository.findById(employeeId).get();
+        updates.forEach((field, value) -> {
+            Field fieldToBeUpdated = ReflectionUtils.findField(EmployeeEntity.class, field);
+            fieldToBeUpdated.setAccessible(true);
+            ReflectionUtils.setField(fieldToBeUpdated, employeeEntity, value);
+        });
+
+        return modelMapper.map(employeeRepository.save(employeeEntity), EmployeeDTO.class);
     }
 }
